restructure stories
